[Bronze V] Title: 이상한 기호, Time: 112 ms, Memory: 12836 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -11,8 +11,8 @@ public class Main {
 		sc.close();
 	}
 	
-	static int oper(int a, int b) {
-		return (a + b) * (a - b);
+	static long oper(int a, int b) {
+		return (long) (a + b) * (a - b);
 	}
 
 }
